fix(util): 时间常量乘法与年份减法显式转为 long，防止将来改动溢出 (sonar S2184)

+++ src/main/java/io/github/jiangood/openadmin/util/datetime/DateTool.java
@@ -105,7 +105,7 @@ public class DateTool {
 
         // 年
         if (begin.length() == 4) {
-            return Integer.parseInt(end) - Integer.parseInt(begin) + 1;
+            return (long) Integer.parseInt(end) - Integer.parseInt(begin) + 1;
         }
         // 年月日
         if (begin.length() == 10) {
